# internal/adapters/postgresql/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package repo

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type TodoPriority string

const (
	TodoPriorityLow    TodoPriority = "low"
	TodoPriorityMedium TodoPriority = "medium"
	TodoPriorityHigh   TodoPriority = "high"
)

func (e *TodoPriority) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TodoPriority(s)
	case string:
		*e = TodoPriority(s)
	default:
		return fmt.Errorf("unsupported scan type for TodoPriority: %T", src)
	}
	return nil
}

type NullTodoPriority struct {
	TodoPriority TodoPriority `json:"todo_priority"`
	Valid        bool         `json:"valid"` // Valid is true if TodoPriority is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTodoPriority) Scan(value interface{}) error {
	if value == nil {
		ns.TodoPriority, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TodoPriority.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTodoPriority) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TodoPriority), nil
}

type TodoStatus string

const (
	TodoStatusTodo       TodoStatus = "todo"
	TodoStatusInProgress TodoStatus = "in_progress"
	TodoStatusCompleted  TodoStatus = "completed"
)

func (e *TodoStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TodoStatus(s)
	case string:
		*e = TodoStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for TodoStatus: %T", src)
	}
	return nil
}

type NullTodoStatus struct {
	TodoStatus TodoStatus `json:"todo_status"`
	Valid      bool       `json:"valid"` // Valid is true if TodoStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTodoStatus) Scan(value interface{}) error {
	if value == nil {
		ns.TodoStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TodoStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTodoStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TodoStatus), nil
}

type Todo struct {
	ID          pgtype.UUID        `json:"id"`
	UserID      pgtype.UUID        `json:"user_id"`
	Title       string             `json:"title"`
	Description pgtype.Text        `json:"description"`
	Status      TodoStatus         `json:"status"`
	Priority    TodoPriority       `json:"priority"`
	DueAt       pgtype.Timestamptz `json:"due_at"`
	CompletedAt pgtype.Timestamptz `json:"completed_at"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
	DeletedAt   pgtype.Timestamptz `json:"deleted_at"`
}

type User struct {
	ID           pgtype.UUID        `json:"id"`
	Email        string             `json:"email"`
	PasswordHash string             `json:"password_hash"`
	DisplayName  pgtype.Text        `json:"display_name"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
	DeletedAt    pgtype.Timestamptz `json:"deleted_at"`
}